Adding HierarchicalExperiment execution in ExperimentBNLauncher and Cleaning Code

+++ src/main/java/org/albacete/simd/clustering/RandomClustering.java
@@ -13,7 +13,7 @@ public class RandomClustering extends Clustering{
         this(42);
     }
 
-    public RandomClustering( int seed){
+    public RandomClustering( long seed){
         Utils.setSeed(seed);
     }
